LTDC demo: Add utility version of LCD code
This has a couple of touchups in the ltdc.c code and
adds the lcd.c file to the utility directory. Now by
including it you get the lcd for your demo. The API
is lcd_flip(int wait_for_TE), and lcd_clear(RGB color)
it gets initialized automatically if you include
util/lcd.o in your makefile with util/retarget.o