Adds some DFT love
So this basically has a very slow DFT calculator (will be an FFT one
here shortly so 100x faster) to plot on the graph. It isn't "accurate"
for reasons that I've not figured out yet, it seems to go 0 to 512hz
instead of 0 - 4096hz which has something to do with the basis functions
I'm pretty sure.