More experiements
Added code to process the bits, collected a sample that should
have a dominant 440hz tone but it wasn't found by the code. So
I'm still perplexed. By toggling PB12 into the WS bit (jumper
pin 11 to 12 on the EXT connector) it seems to reliably start
up the I2S peripheral.